+++ dart/lib/sentry_io.dart
@@ -1,0 +1,2 @@
+export 'sentry.dart';
+export 'src/sentry_attachment/io_sentry_attachment.dart';