+++ dart/lib/src/sentry_envelope.dart
@@ -1,9 +1,11 @@
 import 'dart:convert';
 
+import 'sentry_attachment/sentry_attachment.dart';
 import 'sentry_envelope_header.dart';
 import 'sentry_envelope_item.dart';
 import 'protocol/sentry_event.dart';
 import 'protocol/sdk_version.dart';
+import 'sentry_user_feedback.dart';
 
 /// Class representation of `Envelope` file.
 class SentryEnvelope {
@@ -16,9 +18,29 @@ class SentryEnvelope {
   final List<SentryEnvelopeItem> items;
 
   /// Create an `SentryEnvelope` with containing one `SentryEnvelopeItem` which holds the `SentyEvent` data.
-  factory SentryEnvelope.fromEvent(SentryEvent event, SdkVersion sdkVersion) {
-    return SentryEnvelope(SentryEnvelopeHeader(event.eventId, sdkVersion),
-        [SentryEnvelopeItem.fromEvent(event)]);
+  factory SentryEnvelope.fromEvent(
+    SentryEvent event,
+    SdkVersion sdkVersion, {
+    List<SentryAttachment>? attachments,
+  }) {
+    return SentryEnvelope(
+      SentryEnvelopeHeader(event.eventId, sdkVersion),
+      [
+        SentryEnvelopeItem.fromEvent(event),
+        if (attachments != null)
+          ...attachments.map((e) => SentryEnvelopeItem.fromAttachment(e))
+      ],
+    );
+  }
+
+  factory SentryEnvelope.fromUserFeedback(
+    SentryUserFeedback feedback,
+    SdkVersion sdkVersion,
+  ) {
+    return SentryEnvelope(
+      SentryEnvelopeHeader(feedback.eventId, sdkVersion),
+      [SentryEnvelopeItem.fromUserFeedback(feedback)],
+    );
   }
 
   /// Stream binary data representation of `Envelope` file encoded.
@@ -26,9 +48,10 @@ class SentryEnvelope {
     yield utf8.encode(jsonEncode(header.toJson()));
     final newLineData = utf8.encode('\n');
     for (final item in items) {
-      yield newLineData;
-      await for (final chunk in item.envelopeItemStream()) {
-        yield chunk;
+      final itemStream = await item.envelopeItemStream();
+      if (itemStream.isNotEmpty) {
+        yield newLineData;
+        yield itemStream;
       }
     }
   }