+++ dart/lib/src/sentry_options.dart
@@ -14,6 +14,7 @@ import 'transport/transport.dart';
 import 'utils.dart';
 import 'version.dart';
 import 'platform_checker.dart';
+import 'http_client/sentry_http_client.dart';
 
 // TODO: Scope observers, enableScopeSync
 // TODO: shutdownTimeout, flushTimeoutMillis
@@ -189,6 +190,25 @@ class SentryOptions {
   /// Whether to send personal identifiable information along with events
   bool sendDefaultPii = false;
 
+  /// Whether [SentryEvent] deduplication is enabled.
+  /// Can be further configured with [maxDeduplicationItems].
+  /// Shoud be set to true if
+  /// [SentryHttpClient] is used to capture failed requests.
+  bool enableDeduplication = true;
+
+  int _maxDeduplicationItems = 5;
+
+  /// Describes how many exceptions are kept to be checked for deduplication.
+  /// This should be a small positiv integer in order to keep deduplication
+  /// performant.
+  /// Is only in effect if [enableDeduplication] is set to true.
+  int get maxDeduplicationItems => _maxDeduplicationItems;
+
+  set maxDeduplicationItems(int count) {
+    assert(count > 0);
+    _maxDeduplicationItems = count;
+  }
+
   SentryOptions({this.dsn, PlatformChecker? checker}) {
     if (checker != null) {
       platformChecker = checker;
@@ -250,24 +270,43 @@ typedef BeforeSendCallback = FutureOr<SentryEvent?> Function(
 
 /// This function is called with an SDK specific breadcrumb object before the breadcrumb is added
 /// to the scope. When nothing is returned from the function, the breadcrumb is dropped
-typedef BeforeBreadcrumbCallback = Breadcrumb? Function(Breadcrumb? breadcrumb,
-    {dynamic hint});
-
-/// Logger interface to log useful debugging information if debug is enabled
-typedef SentryLogger = void Function(SentryLevel level, String message);
+typedef BeforeBreadcrumbCallback = Breadcrumb? Function(
+  Breadcrumb? breadcrumb, {
+  dynamic hint,
+});
 
 /// Used to provide timestamp for logging.
 typedef ClockProvider = DateTime Function();
 
+/// Logger interface to log useful debugging information if debug is enabled
+typedef SentryLogger = void Function(
+  SentryLevel level,
+  String message, {
+  Object? exception,
+  StackTrace? stackTrace,
+});
+
 /// A NoOp logger that does nothing
-void noOpLogger(SentryLevel level, String message) {}
+void noOpLogger(
+  SentryLevel level,
+  String message, {
+  Object? exception,
+  StackTrace? stackTrace,
+}) {}
 
 /// A Logger that prints out the level and message
-void dartLogger(SentryLevel level, String message) {
+void dartLogger(
+  SentryLevel level,
+  String message, {
+  Object? exception,
+  StackTrace? stackTrace,
+}) {
   log(
     '[${level.name}] $message',
     level: level.toDartLogLevel(),
     name: 'sentry',
     time: getUtcDateTime(),
+    error: exception,
+    stackTrace: stackTrace,
   );
 }